restructered format of obstacle environments, and created seperate run class. Only updated in single agent environment area so far
