+++ MultiAgentEnvironment.py
@@ -354,6 +354,6 @@ def run_scenario_3():
 
 
 # Run the scenarios
-run_scenario_1()
+#run_scenario_1()
 #run_scenario_2()
 #run_scenario_3()
